The WId can be a HWD on Windows, so explicitly cast to make that work. urgh

+++ agents/notesagent/notesagent.cpp
@@ -93,7 +93,7 @@ bool NotesAgent::enabledAgent() const
 
 void NotesAgent::configure( WId windowId )
 {
-    showConfigureDialog(windowId);
+    showConfigureDialog((qlonglong)windowId);
 }
 
 void NotesAgent::showConfigureDialog(qlonglong windowId)
